Adjust class and namespace naming
