+++ Logger.cs
@@ -3,7 +3,7 @@ using System.IO;
 using System.Text;
 using System.Text.RegularExpressions;
 
-namespace MusicBeePlugin
+namespace MPrisBee
 {
     public class Logger
     {
